unet_project/requirements.txt
numpy
pandas
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow>=1.3.0
keras>=2.0.8
opencv-python
h5py
imgaug
scikit-image
tqdm
joblib
IPython[all]


